refactor(bridge): move submit/status failure classifiers to status-controller
Keep quote-fetch classification in bridge-controller. Hash presence and submit/status helpers belong with the controller that will emit them.
