+++ packages/bridge-controller/CHANGELOG.md
@@ -9,10 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add SwapBridge failure telemetry enums and classifiers for a later emit ([#9947](https://github.com/MetaMask/core/pull/9947))
-  - New exports: `FailurePhase`, `SwapBridgeErrorCode`, and helpers that classify quote, submit, and status failures from the code path (not from `error_message`)
+- Add SwapBridge failure telemetry schema and quote-fetch classifier for a later emit ([#9947](https://github.com/MetaMask/core/pull/9947))
+  - New exports: `FailurePhase`, `SwapBridgeErrorCode`, and `getQuoteFetchErrorCode` (classifies quote fetch from the code path, not from `error_message`)
   - Optional `failure_phase` and `error_code` on Quotes Error and Failed event context types
   - Optional `source_hash_present` and `destination_hash_present` on Failed, Submitted, and Completed event context types
+  - Submit and status classifiers live in `@metamask/bridge-status-controller`
 
 ### Changed
